requirements.txt
flask
flask-cors
google-generative-ai
python-dotenv
werkzeug
